backend/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
pydantic==2.7.4
gitpython==3.1.43
radon==6.0.1
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.2